Project02.02
I want to make clear, this isn't a Hardware description language. It's analogous to a hardware description language. A real HDL would functionality supporting logic things like loops or some sort of statement, all this possesses is the ability to declare items and this may look fucking ugly and like a crime against god but this is the best this language can do. Just Hardcode after Hardcode after Hardcode. It is not my fault, do not judge me based off of this just be amazed that it works.